display ref used in benchmark

+++ pull.sh
@@ -4,6 +4,11 @@
 
 set -e
 
+if [ -z "$1" ] || [ -z "$2" ]; then
+  echo "Expected to receive two refs. Cannot continue."
+  exit 1
+fi
+
 rm -rf underscore
 rm -f vendor/*.js
 
@@ -13,9 +18,11 @@ cd underscore
 
 git checkout $1
 cp underscore.js ./../vendor/underscore-baseline.js
+echo "{\"ref\":\"$1\"}" > ./../vendor/underscore-baseline.ref.json
 
 git checkout $2
 cp underscore.js ./../vendor/underscore-comparison.js
+echo "{\"ref\":\"$2\"}" > ./../vendor/underscore-comparison.ref.json
 
 cd ..
 rm -rf underscore
